+++ src/bp/admin/ui/src/auth/LoginContainer.tsx
@@ -1,8 +1,10 @@
+import { Callout, Intent } from '@blueprintjs/core'
 import { lang } from 'botpress/shared'
+import cx from 'classnames'
 import React, { FC } from 'react'
-import { Alert, Card, CardBody, CardText, CardTitle } from 'reactstrap'
 
 import logo from './media/nobg_white.png'
+import style from './style.scss'
 
 interface Props {
   title?: string
@@ -14,24 +16,26 @@ interface Props {
 
 const LoginContainer: FC<Props> = props => {
   return (
-    <div className="centered-container">
-      <div className="middle">
-        <div className="inner">
-          <img className="logo" src={logo} alt="loading" />
-          <Card body>
-            <CardBody className="login-box">
+    <div className={cx('centered-container', style.centered_container)}>
+      <div className={cx('middle', style.middle)}>
+        <div className={cx('inner', style.inner)}>
+          <img className={cx('logo', style.logo)} src={logo} alt="loading" />
+          <div className={cx('card', 'card_body', style.card, style.card_body)}>
+            <div className={cx('card_body', 'login_box', style.card_body, style.login_box)}>
               <div>
-                <CardTitle>
+                <div className={cx('card_title', style.card_title)}>
                   <strong>{props.title || 'Botpress'}</strong>
-                </CardTitle>
-                <CardText>{props.subtitle || ''}</CardText>
-                {props.error && <Alert color="danger">{props.error}</Alert>}
+                </div>
+
+                <div className={cx('card_text', style.card_text)}>{props.subtitle || ''}</div>
+
+                {props.error && <Callout intent={Intent.DANGER}>{props.error}</Callout>}
                 {props.children}
               </div>
-            </CardBody>
-          </Card>
+            </div>
+          </div>
           {props.poweredBy && (
-            <div className="homepage">
+            <div className={cx('homepage', style.homepage)}>
               <p>
                 {lang.tr('admin.poweredBy')} <a href="https://botpress.com">Botpress</a>
               </p>